Refactor wellness settings into dedicated view file
